1.5.0-0a
* Updated to 1.21.11

+++ ref-decompiled/net/minecraft/block/vault/VaultClientData.java
@@ -1,25 +1,33 @@
+/*
+ * Decompiled with CFR 0.152.
+ * 
+ * Could not load the following classes:
+ *  net.minecraft.block.vault.VaultClientData
+ *  net.minecraft.util.math.MathHelper
+ */
 package net.minecraft.block.vault;
 
 import net.minecraft.util.math.MathHelper;
 
 public class VaultClientData {
-   public static final float DISPLAY_ROTATION_SPEED = 10.0F;
-   private float displayRotation;
-   private float lastDisplayRotation;
+    public static final float DISPLAY_ROTATION_SPEED = 10.0f;
+    private float displayRotation;
+    private float lastDisplayRotation;
 
-   VaultClientData() {
-   }
+    VaultClientData() {
+    }
 
-   public float getDisplayRotation() {
-      return this.displayRotation;
-   }
+    public float getDisplayRotation() {
+        return this.displayRotation;
+    }
 
-   public float getLastDisplayRotation() {
-      return this.lastDisplayRotation;
-   }
+    public float getLastDisplayRotation() {
+        return this.lastDisplayRotation;
+    }
 
-   void rotateDisplay() {
-      this.lastDisplayRotation = this.displayRotation;
-      this.displayRotation = MathHelper.wrapDegrees(this.displayRotation + 10.0F);
-   }
+    void rotateDisplay() {
+        this.lastDisplayRotation = this.displayRotation;
+        this.displayRotation = MathHelper.wrapDegrees((float)(this.displayRotation + 10.0f));
+    }
 }
+
